backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
pillow
python-dotenv
pydantic
requests
numpy
python-dotenv
torch
torchvision
